add a base workflow to embed from it all the workflow concretions
